Added KickHeartbeat
Server sends two heartbeats, one for connecting, the other to kick the
player and tell them to connect on the other heartbeat. Example: Server
sends normal heartbeat to classicube.net, Server also sends a
KickHeartbeat to minecraft.net. If a player joins on minecraft.net it
kicks them and tells them to join on "classicube.net"